+++ bgpy/cli.py
@@ -1,6 +1,7 @@
 from .server import Server
 from .client import Client
-from typer import Typer, echo, Abort, Argument, Option
+from .core.token import token_getenv, token_setenv
+from typer import Typer, echo, Abort, Argument, Option, prompt
 from typing import Optional
 from pathlib import Path
 
@@ -16,6 +17,12 @@ app = Typer(add_completion=False)
 def run_server(
     host: str = Argument(..., help="Host address to run the server on"),
     port: int = Argument(..., help="Port where the server should listen to"),
+    token: bool = Option(
+        False,
+        "--token",
+        "-t",
+        help="Prompt for a token for the client server communication",
+    ),
     log_level: str = Option(
         "INFO",
         "--log-level",
@@ -24,6 +31,15 @@ def run_server(
     ),
     log_file: Optional[Path] = Option(
         None, "--log-file", "-f", help="Path to a log file"
+    ),
+    init_file: Optional[Path] = Option(
+        None,
+        "--init-file",
+        "-i",
+        help=(
+            "Path to a file containing the 'init_task', 'exec_task'"
+            + "and 'exit_task' for the server initialization"
+        ),
     ),
 ) -> None:
     """
@@ -35,10 +51,19 @@ def run_server(
     passing 'init_task()', exec_task()' and 'exit_task()' to the server, the
     server will not respond to requests.
     """
+    if token:
+        TOKEN = prompt("Token", hide_input=True)
+        token_setenv(TOKEN)
     if str(log_file) == "None":
         log_file = None
+    if str(init_file) == "None":
+        init_file = None
     server = Server(
-        host=host, port=int(port), log_level=log_level, log_file=log_file
+        host=host,
+        port=int(port),
+        log_level=log_level,
+        log_file=log_file,
+        init_file=init_file,
     )
     try:
         server.run()
@@ -51,6 +76,12 @@ def run_server(
 def terminate_server(
     host: str = Argument(..., help="Host address of the server"),
     port: int = Argument(..., help="Port where the server is listening"),
+    token: bool = Option(
+        False,
+        "--token",
+        "-t",
+        help="Prompt for a token for the client server communication",
+    ),
     log_level: str = Option(
         "INFO",
         "--log-level",
@@ -67,10 +98,19 @@ def terminate_server(
     Terminate a bgpy server on the given host, which is listening to the
     provided port.
     """
+    if token:
+        TOKEN = prompt("Token", hide_input=True)
+        token_setenv(TOKEN)
+    else:
+        TOKEN = token_getenv()
     if str(log_file) == "None":
         log_file = None
     client = Client(
-        host=host, port=int(port), log_level=log_level, log_file=log_file
+        host=host,
+        port=int(port),
+        token=TOKEN,
+        log_level=log_level,
+        log_file=log_file,
     )
     try:
         client.terminate()